+++ InventorAccessPortal.Web/Mailer/Models/ResetPassword/EmailModel.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Web;
 
-namespace InventorAccessPortal.Web.Mailer.Models.CompleteRegistration
+namespace InventorAccessPortal.Web.Mailer.Models.ResetPassword
 {
     public class EmailModel: _EmailModel
     {
