--- InventorAccessPortal.Web/Mailer/Models/ResetPassword/PasswordModel.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-
-namespace InventorAccessPortal.Web.Mailer.Models.ResetPassword
-{
-    public class PasswordModel
-    {
-        public string Email { get; set; }
-        public string Password { get; set; }
-        public string PasswordVerify { get; set; }
-    }
-}
