+++ InventorAccessPortal.Web/Mailer/Models/_EmailModel.cs
@@ -8,6 +8,6 @@ namespace InventorAccessPortal.Web.Mailer.Models
 {
     public class _EmailModel : _ErrorModel
     {
-
+        public Guid ActionGuid;
     }
 }
