+++ InventorAccessPortal.Web/Mailer/Templates/ResetPassword/Received.cshtml
@@ -1,0 +1,36 @@
+﻿@model InventorAccessPortal.Web.Models.Account.ResetPasswordModel
+@using InventorAccessPortal.Web.Models.Account
+@{
+    ViewBag.Title = "ResetPassword";
+    Layout = "~/Views/Shared/_Layout.cshtml";
+}
+<h1>Reset Password</h1>
+@using (Html.BeginForm("ActualResetPassword", "Account", FormMethod.Post))
+{
+    @Html.AntiForgeryToken()
+    <div class="form-group row">
+        <label for="email" class="col-sm-2 col-form-label">Email:</label>
+        <div class="col-sm-10">
+            @Html.TextBoxFor(p => p.Email, new { id = "email", disabled = "true" })
+        </div>
+    </div>
+    <div class="form-group row">
+        <label for="password" class="col-sm-2 col-form-label">Password:</label>
+        <div class="col-sm-10">
+           @Html.Password("password", new { id = "password", @class = "verifyThisPassword", data_verifier = "#passwordVerification" })
+            <div class="form-text text-muted">
+                <small>Passwords length must be 16-255 characters</small>
+            </div>
+        </div>
+    </div>
+    <div class="form-group row">
+        <label for="passwordVerification" class="col-sm-2 col-form-label">Verify Password:</label>
+        <div class="col-sm-10">
+            @Html.Password("passwordVerification", new { id = "passwordVerification" })
+        </div>
+    </div>
+
+    <input class="btn btn-primary" type="submit" value="Reset Password" />
+    @Html.Partial("~/Views/Shared/_IgnoreFormWarnings.cshtml")
+    @Scripts.Render("~/bundles/account/scripts")
+}
