+++ InventorAccessPortal.Web/Mailer/Templates/ResetPassword/ResetPasswordError.cshtml
@@ -3,5 +3,5 @@
     ViewBag.Title = "Reset Password Error";
 }
 
-<h2>Error reset your password.</h2>
+<h2>Error reseting your password.</h2>
 <h4>Please try to reset again later.</h4>
