+++ InventorAccessPortal.Web/Views/Account/ResetPassword.cshtml
@@ -5,9 +5,10 @@
     Layout = "~/Views/Shared/_Layout.cshtml";
 }
 <h1>Reset Password</h1>
-@using (Html.BeginForm("ActualResetPassword", "ResetPassword", FormMethod.Post))
+@using (Html.BeginForm("ResetPassword", "Account", FormMethod.Post))
 {
     @Html.AntiForgeryToken()
+
     <div class="form-group row">
         <label for="email" class="col-sm-2 col-form-label">Email:</label>
         <div class="col-sm-10">
@@ -19,23 +20,9 @@
             </div>
         </div>
     </div>
-    <div class="form-group row">
-        <label for="password" class="col-sm-2 col-form-label">Password:</label>
-        <div class="col-sm-10">
-           @Html.PasswordFor(p => p.Password, new { id = "password", value = Model.Password, @class = "verifyThisPassword", data_verifier = "#passwordVerification" })
-            <div class="form-text text-muted">
-                <small>Passwords length must be 16-255 characters</small>
-            </div>
-        </div>
-    </div>
-    <div class="form-group row">
-        <label for="passwordVerification" class="col-sm-2 col-form-label">Verify Password:</label>
-        <div class="col-sm-10">
-            @Html.PasswordFor(p => p.PasswordVerify, new { value = Model.PasswordVerify, id = "passwordVerification" })
-        </div>
-    </div>
 
     <input class="btn btn-primary" type="submit" value="Reset Password" />
     @Html.Partial("~/Views/Shared/_IgnoreFormWarnings.cshtml")
     @Scripts.Render("~/bundles/account/scripts")
 }
+
