--- InventorAccessPortal.Web/Views/ResetPassword/ActualResetPassword.cshtml
@@ -1,17 +1,0 @@
-﻿
-@{
-    Layout = null;
-}
-
-<!DOCTYPE html>
-
-<html>
-<head>
-    <meta name="viewport" content="width=device-width" />
-    <title>ActualResetPassword</title>
-</head>
-<body>
-    <div> 
-    </div>
-</body>
-</html>
