--- InventorAccessPortal.Web/Views/ResetPassword/ResetPassword.cshtml
@@ -1,28 +1,0 @@
-﻿@model InventorAccessPortal.Web.Models.Account.ResetPasswordModel
-@using InventorAccessPortal.Web.Models.Account
-@{
-    ViewBag.Title = "ResetPassword";
-    Layout = "~/Views/Shared/_Layout.cshtml";
-}
-<h1>Reset Password</h1>
-@using (Html.BeginForm("ResetPassword", "ResetPassword", FormMethod.Post))
-{
-    @Html.AntiForgeryToken()
-
-    <div class="form-group row">
-        <label for="email" class="col-sm-2 col-form-label">Email:</label>
-        <div class="col-sm-10">
-            @Html.TextBoxFor(p => p.Email, new { id = "email" })
-            <div class="form-text text-muted">
-                <small>
-                    Email must be in format: user@domain.(com, edu, net, etc.) and less than 255 characters
-                </small>
-            </div>
-        </div>
-    </div>
-
-    <input class="btn btn-primary" type="submit" value="Reset Password" />
-    @Html.Partial("~/Views/Shared/_IgnoreFormWarnings.cshtml")
-    @Scripts.Render("~/bundles/account/scripts")
-}
-
