+++ InventorAccessPortal.Web/Views/Shared/_SidebarPartial.cshtml
@@ -61,8 +61,8 @@
                     </a>
                 </li>
                 <li>
-                    <a href="@Url.Action("ResetPassword", "ResetPassword")" class="btn btn-primary">
-                        Reset Passwrod
+                    <a href="@Url.Action("ResetPassword", "Account")" class="btn btn-primary">
+                        Reset Password
                         <span class="icon glyphicon glyphicon-plus-sign" aria-hidden="true"></span>
                     </a>
                 </li>
